assignments/gmail_client/requirements.txt
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
MySQL-python==1.2.5
mysqlclient==1.3.12
oauth2client==4.1.2
python-dateutil==2.7.3
